server/main.py
from fastapi import FastAPI, UploadFile, File, BackgroundTasks, Depends
from fastapi.middleware.cors import CORSMiddleware
import os
from pydantic import BaseModel
from typing import List, Optional
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(title="LectaAI API")

# CORS configuration
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Models
class ChatMessage(BaseModel):
    lecture_id: str
    message: str

class UserProfile(BaseModel):
    id: str
    email: str
    role: str

@app.get("/")
async def root():
    return {"message": "LectaAI API is running"}

@app.post("/upload")
async def upload_lecture(
    background_tasks: BackgroundTasks,
    file: UploadFile = File(...),
    user_id: str = None
):
    # Logic for:
    # 1. Saving file to storage
    # 2. Triggering Whisper transcription in background
    # 3. Generating notes via LLM
    return {"status": "processing", "lecture_id": "mock_id_123"}

@app.post("/chat")
async def chat_with_lecture(chat: ChatMessage):
    # Logic for:
    # 1. Retrieving lecture context (RAG)
    # 2. Querying LLM
    return {"response": "This is an AI generated response based on the lecture context."}

@app.get("/lectures/{lecture_id}")
async def get_lecture_data(lecture_id: str):
    return {
        "id": lecture_id,
        "title": "Introduction to Quantum Physics",
        "transcript": "...",
        "notes": "...",
        "summary": "...",
        "quizzes": []
    }

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
