server/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
supabase
pydantic
slowapi
pydub
numpy
faiss-cpu
